requirements.txt
astroid==2.15.5
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
dill==0.3.6
google-auth==2.21.0
google-auth-oauthlib==1.0.0
gspread==5.9.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.3
platformdirs==3.8.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pylint==2.17.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomlkit==0.11.8
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
